voice_clone_demo/requirements.txt
torch
torchaudio
transformers
gradio
scipy
einops
tqdm
tortoise-tts @ git+https://github.com/neonbjb/tortoise-tts.git